Moved split_primitives_at_intersection to split
